set up separate files for velocity and libseis
