+++ src/velocity.h
@@ -1,0 +1,16 @@
+//
+// Created by dg on 5/4/24.
+//
+
+#ifndef VELOCITY_H
+#define VELOCITY_H
+
+typedef struct {
+  int rows;
+  int cols;
+  double **data;
+} VelocityModel;
+
+VelocityModel *allocate_velocity_model(int rows, int cols);
+
+#endif // VELOCITY_H
